pyproject.toml
[project]
name = "anamnesa"
version = "0.0.1"
description = "Indonesian clinical guideline retrieval agent"
requires-python = ">=3.12"
authors = [{ name = "Rifqi" }]
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.8",
    "structlog>=24.1",
    "python-ulid>=3.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    # retrieval
    "lancedb>=0.14",
    "rank-bm25>=0.2.2",
    "mcp>=1.2",
    # ingestion
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    # web / SSE
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.11",
]
# Real multilingual embedder — pulls in torch, ~2GB download of BGE-M3
# weights on first use. Only install when you're ready to reindex.
#     uv pip install -e '.[embeddings]'
embeddings = [
    "sentence-transformers>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "mcp", "scripts", "tools", "server"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "A", "ASYNC", "RUF"]
ignore = []

[tool.ruff.lint.per-file-ignores]
# Deploy / bench / probe scripts intentionally do:
#   1. load_dotenv() before importing heavyweight modules that read env
#      (ANTHROPIC_API_KEY, ANAMNESA_EMBEDDER, etc.) at import time → E402;
#   2. emit long curl / shell-quoted payloads that'd be worse folded → E501.
# These patterns aren't bugs — suppress at the directory level.
"scripts/*.py" = ["E402", "E501"]
# Video narration scripts contain complete sentences as string literals
# (for ElevenLabs TTS). Folding them onto multiple lines hurts readability
# and makes it harder to see what the voice actually reads.
"video/*.py" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
